+++ app/src/main/java/tk/only5/ptsmoodle/LoginSignUpFragmentActivity.java
@@ -2,13 +2,7 @@ package tk.only5.ptsmoodle;
 
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
-import android.util.Log;
 import android.view.WindowManager;
-
-import com.parse.ParseException;
-import com.parse.ParseInstallation;
-import com.parse.ParsePush;
-import com.parse.SaveCallback;
 
 public class LoginSignUpFragmentActivity extends AppCompatActivity {
 
@@ -21,22 +15,6 @@ public class LoginSignUpFragmentActivity extends AppCompatActivity {
         getWindow().setSoftInputMode(
                 WindowManager.LayoutParams.SOFT_INPUT_STATE_ALWAYS_HIDDEN
         );
-        ParseInstallation.getCurrentInstallation().saveInBackground(new SaveCallback() {
-            @Override
-            public void done(ParseException e) {
-                Log.d(TAG, ParseInstallation.getCurrentInstallation().getString("deviceToken") + "");
-                if (e == null) {
-                    Log.d(TAG, "Installation Saved");
-                    ParsePush.subscribeInBackground("TEST", new SaveCallback() {
-                        @Override
-                        public void done(ParseException e) {
-                            if (e == null) Log.d(TAG, "Channel Subscribed");
-                            else Log.e(TAG, "ERROR", e);
-                        }
-                    });
-                } else Log.e(TAG, "ERROR", e);
-            }
-        });
         Functions.setFragment(getSupportFragmentManager(), new LoginFragment(), getResources().getString(R.string.label_login), R.id.fragmentContainerLoginSignUp, false);
 
     }
